Module Oidc

High level API

module SimpleClient : sig ... end

Simpler interface for creating a oidc client

Lower level constructs

module Client : sig ... end

Types and functions to work with clients

module Discover : sig ... end

Parsing and creating the discovery document. All fields listed here: https://openid.net/specs/openid-connect-discovery-1_0.html#ProviderMetadata

module Parameters : sig ... end

Auth parameters

module Scopes : sig ... end

https://openid.net/specs/openid-connect-basic-1_0.html#Scopes

module Token : sig ... end

Types and functions to work with the token endpoint

module IDToken : sig ... end

ID Token validation and creation

module Jwks : sig ... end

JSON Web Keys

module Userinfo : sig ... end

module Pkce : sig ... end

* Copyright 2022 Ulrik Strid. All rights reserved. * Use of this source code is governed by a BSD-style * license that can be found in the LICENSE file.

module Error : sig ... end