Module Oidc.Token

Types and functions to work with the token endpoint

OpenID Connect Core 1.0 TokenEndpoint

module Response : sig ... end

module Request : sig ... end

Types and functions to work with the token endpoint

Utils

val basic_auth : client_id:string -> secret:string -> string * string

Creates a valid Basic auth header from client_id and secret